lxy-merge

+++ common/precomp.h
@@ -19,16 +19,16 @@ std::string myPlaneImageFile = ":/images/myplane.png";
 std::string bossImageFile = ":/images/boss2.gif";//boss图片
 
 QPixmap pixmap1(QPixmap(QString::fromStdString(bossImageFile)));
-const int bosswidth= pixmap1.width();
-const int bossheight= pixmap1.height();
+const double bosswidth= pixmap1.width();
+const double bossheight= pixmap1.height();
 
 QPixmap pixmap2(QPixmap(QString::fromStdString(enemyPlaneImageFile)));
-const int enemyplanewidth= pixmap2.width();
-const int enemyplaneheight= pixmap2.height();
+const double enemyplanewidth= pixmap2.width();
+const double enemyplaneheight= pixmap2.height();
 
 QPixmap pixmap3(QPixmap(QString::fromStdString(myPlaneImageFile)));
-const int myplanewidth= pixmap3.width();
-const int myplaneheight= pixmap3.height();
+const double myplanewidth= pixmap3.width();
+const double myplaneheight= pixmap3.height();
 
 typedef double POS;
 typedef std::vector<POS*> POSES;
